보스 시스템

+++ Assets/3.Scripts/Interface/IGetDamage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IGetDamage
 {
-    void GetDamage(int damage);
+    void GetDamage(float damage);
 }
 
 public interface IGetPlayerDamage : IGetDamage
